Back completo

+++ controller/user/deleteUser.php
@@ -1,6 +1,8 @@
 <?php
 
 include_once "cors.php";
+include "../../JWTValidate.php";
+
 include "../../model/UserModel.php";
 include "../../DAO/DAOUser/deleteUser.php";
 include "../../DAO/DAOUser/gettingType.php";
@@ -10,7 +12,15 @@ $getData = json_decode(file_get_contents("php://input"));
 
 $adminFlag = isAdmin($getData);
 
-if($adminFlag){
+$JWTCheck = "init";
+
+try {
+    $JWTCheck = Auth::Check($getData->token);
+} catch (\Throwable $th) {
+    
+}
+
+if($adminFlag && $JWTCheck==null){
 
     $resultDao = deleteUser($getData->userToDelete);
 
@@ -26,6 +36,12 @@ if($adminFlag){
 
         echo json_encode($data);
     }
+
+} else if($JWTCheck!=null){
+
+    $data = array("ok" => "false", "msg" => "Expired");
+
+    echo json_encode($data);
 
 } else {
